requirements.txt
asgiref==3.8.1
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
XlsxWriter==3.2.0
